Fix an issue with ElxResizeObserver
If `observe` was called on the same element twice, the service would
previously close any existing streams for that element before creating
a new one. Not only was this a little wasteful, but it also caused some
unexpected behavior for consumers, with the older subscriber ceasing to
receive notifications without any warning.

This has been updated to simply return the existing stream if it
already exists and is not closed. This could potentially cause some
further unexpected behavior, since the `options` argument passed by
late subscribers will be ignored. That would require more effort to
fix than I'm willing to invest at the moment, so in this case we'll
log a warning so consumers aren't completely in the dark, and cross
that bridge properly once there's a need for it.

In the meantime, if it's required to observe the same element from
two different consumers with two different sets of options, the second
consumer should be able to provide its own unique instance of the
service instead of injecting an existing one.

We also add a check to `unobserve` to ensure that the stream has no
remaining observers before closing and deleting it.